Improve termination conditions
Make them more generic (we begin to see an interface there).
Add the 'started' status, and count time-before-timeout from this starting
point, instead of from the beginning of the simulation.
DominoRunTerminationCondition uses a queue of dominoes emptied as they topple,
instead of an index (more concise).
AndTerminationCondition correctly updates its status and timeouts if a branch
does not start at all.